internal/rest/access: Move Microcluster specific auth bits to internal package
